Normalise single-plist text property form
When trying engrave-faces in an Org buffer I noticed that sometimes it
seems a single plist can be returned, e.g. (:inherit org-block). This is
problematic as the processing method requires a list where each list
item is a text property form, which does not hold true in this case.

We can correct for this by checking to see if the obtained text-property
has the form (:keyword ...), and if so wrap it in a list.